First draft (Albums and Artists)
After struggling with AppSync and DynamoDB for a while I finally managed
to wrap my head around it and arrive to a nice solution for the Artists
and Albums entities. Now that I understand better the constraints of
NoSQL and how to work with it, it's time to finish the exploration
phase, merge the current state of the client into main and use this
experience as foundation for the rest.