Trying out many ideas
This was a months-long branch in which I tried many different ideas. The
code went through many refactorings and rewrites as I was experimenting
with what's possible, and ran into roadblocks that forced me to review
my previous assumptions.

It started with simply adding the shiva indexer to the repo, but quickly
moved to adapting the DB schema to the new data, displaying that data in
the UI, finding and correcting bugs along the way, and even rewriting
full parts of the app because the new data challenged the current state
of things.

I did not want to commit the many experimental commits that I had in
this branch nor split all the code into atomic comits because it made no
sense to put so much effort into code that was written and rewritten and
rewritten again in successive commits. So I'm just squashing it all
together here trying to finally drop this way-too-long-lived branch.