Convert queue from route into floaty-thingy
The queue will not be a route any more because it's not a
shareable-resource. It makes no sense to, for example, copy the URL of
the route and share it with someone, so it should not have a URL at all.

Instead it's now part of the app's state, it will be always rendered but
hidden, and it will slide up on request without triggering a navigation.

The only possible problem is that users may be tempted to hit the back
button to exit the queue. I'm not sure it's possible to intercept and
close the queue instead of triggering a navigation.