Add a blurred background image to the header
The gradient felt a bit cheap, I think this gives a bit more of an
"app-feeling" to it.

The `backdrop-filter` css directive introduced a very hard to track
bug, messing up the stack order of the elements. As a result, the
options component showed up behind the shoulder. For some details, see:
https://stackoverflow.com/a/57327698