Make ScrollTop component slide in after scrolling
Instead of having to scroll all the way to the bottom to find the
component, now it shows up after scrolling passed 1 viewport of results.
Another approach we tried was to make the header sticky but the 2
possible approaches had their problems:

Leaving the header fixed on top and make only the shoulder scrollable:
We would lose the scroll position after navigating away and back. The
possible solutions are a bit too hacky.

Make the header sticky on scroll: This just doesn't feel as snappy as it
does on native apps. Even worse, with the arch under the header things
are even more complicated. All the research showed that the complexity
was definitely not worth it.

Finally, we also remove the component from the queue because we don't
expect that list to be nearly as extense as an artist or album list.