Remove animation from queue when hidden
In mobile, when you click on the search and the keyboard is displayed,
the viewport height changes. When you dismiss the keyboard the queue had
adapted to the new page height and it will show up where the keyboard
used to be and transition off the screen. That's why we need to remove
the transition when the queue is hidden.