iOS hack: Avoid the dismiss animation altogether
There is no back button on Safari, instead the user navigates back by
swiping from the edge of the screen. When this happens, if we don't
remove the animation, the queue shows up on the previous page and only
then slides away.

Now, at most we get a flash of the queue sometimes but it disappears
immediately. Since "our" back button is covered by the queue there is no
other way of navigating back than swiping, so there is no duality there.