Move all app state to a centralized reducer
The idea here is to gather all non-persisten state spread accross the
app (i.e. whether the user is admin, whether the browser is safari, and
the visibility of the queue) and move it to a centralized place.

For that I'm creating a reducer and making it available to the whole app
through a context. The problem is that the components that need to read
the state can do it only once but they don't get any subsequent updates.

For example, the queue loads, reads the initial state (`queueOpen` is
`false`) and when the player is clicked to open the queue, you can see
that the queue re-renders but it receives the old `queueOpen` value. At
the same time, the reducer has executed and updated the value to `true`,
but the update never reaches the queue, so it doesn't open. It
re-renders in the closed state.

That's why I'm removing leaving that logic in the `root.tsx` file, and
telling the queue whether to render visible or not. In the future we
may need to open the queue from different places in the app, for which
reacting to changes in the state would be useful but for the time being
this is enough.