Remove `.gitlab-ci.yml` file
It was outdated and useless. We need to design a proper pipeline
strategy that considers:
* There are no tests really, but at least we should build it.
* Do we want to deploy a stage?
* Will that stage be ephemeral or long-lived?

There are advatages to both approaches, ephemeral stages are good to
test the whole deployment process each time, but the CDN (and certs)
creation take usually very long (and could even timeout). At the same
time, a developer would not create a new stage each time, so this is a
step that is not executed very often.

On the other hand, having a long-lived stage is much quicker to deploy
and can surface problems with migrations that would go unnoticed in
brand new deployments.