Improved line-breaks in the queue and album detail
In the queue the artist name and the album name would break
independently, now they are a continuation of each other and will break
as if they were just one element. At the same time the cross will always
remain vertically-aligned at the center, regardless of the element's
height.

In the album detail, however, when the space is not enough for the line
it won't break at the end, instead it will leave the artist name in the
first line and move the rest of the details to the second line, without
breaking either of them.