Remove auxiliary variable `_album`
This variable was introduced as a workaround for a bug that forced the
whole app to render on any state change, creating a "loading" flash for
less than a second. To prevent it we would store the album in this
auxiliary variable and only re-render if it changed to any value except
`null`.

The bug was fixed at some point but we still have these helper variables
lying around.