Remove all history manipulation
Since it's not possible to handle all corner cases when it comes to
history manipulation, it's better to not try at all. The code complexity
increases exponentially for a feature that it's not even complete.

The problem is the different behaviour in android and ios when it comes
to navigation. In android the user has a back button, so we wanted to
close the queue when the user presses the back button. However, since
it's not possible to "pop" a state from the history, once the user opens
and closes the queue, he/she would have to press "back" twice (actually,
once per each time the queue was opened and closed consequtively) before
being able to navigate back.

On iOS, however, there is no back button. Instead, the user navigates by
swiping the page "away". In this case the user would go back to the
previous page but with the queue still open and it would then slide
away. For that case the `isSafari` flag was introduced, in order to
disable the animation when a navigation happened in Safari.

So we ended up with a lot of auxiliary variables and conditional code in
order to decide whether to navigate back, or close the queue, but it did
not work in all cases (e.g. the double back issue with Android). And let
alone the forward button, in which we would have to keep the listener at
all times, even with the queue closed, just to detect if the user
navigated forward into an open-queue state again.

If the feature was possible then it would have been worth the effort,
but given that the end result is a half-assed solution with an
unmanageable code complexity, it's better to not even try at all.