Display the invite link as a QR code
In the old days in order to share a record with someone you had to
physically meet them and give them the medium which held the music. To
mimic this experience, instead of sharing a link through a chat app, you
would have to display the code in your phone to the person that you want
to invite.

Of course this does not prevent taking a screenshot and sending it to
someone who is in a remote location, but it's not the point to prevent
that since it would be futile to even try.

Links had issues of their own, for example nowadays most chat apps stick
their noses into every link that comes their way, making it impossible
to share a 1-time link.