Improve authentication step
Namely:
* Add an online/offline check
* Retry on NotAuthenticatedException
* Display error messages
* Add a retry button

When the user is offline we display an appropriate message, the problem
is that the browser detects that the wifi connects but does not perform
any kind of preflight checks, so it fires off the request against
cognito before the connection is fully stablished.

This meant that after going back online we would get an "unauthorized"
message every time, forcing the user to reload. To fix this we retry 3
times after the browser detects that we are back online, and if that
doesn't work then the user can manually retry again (which is just a
`location.reload()` actually).