Fix track highlight
It was looking weird with only some parts of the track title changing
color. Instead we leave the color alone and use the same highlighting
that we use for the queue, i.e. a very light background with darker top
and bottom borders.

This meant removing the shoulder's padding, which applies for all detail
pages, that's why the artist detail had to be adapted as well.