Remove all resources related to Playlists and Favs
We don't need them now and when we implement them we will most likely do
it client-side. It would be almost trivial to implement a
LocalStorage-backed list of favorites or playlists, with the option to
back it up in the server, serialized, in some form of generic user data
field in dynamo. Thanks @medwards for the idea, I think it will simplify
the backend a lot.