Limit uses of `getClient` to `use` and `useGet`
Several files all over the place were calling `getClient` directly which
made it challenging to add support for multiple servers. At some point
in the code we have to decide which server to query and having the
client spread all over the codebase meant that we had to make that same
decision in multiple places.

Now it's only in 2 places, but they both make sense since they are the 2
points that make either a query to get all elements or 1. We can
centralize here the decision of which server to query, and in the future
we could even unify those 2 into 1.