Turn errors into strings
When we treat errors as objects we have to make some special checks to
avoid infinite loops. When you call `setError` preact will do a `===`
check to decide whether to re-render, and every time we get an error,
even if it's the same kind of error, will always be a different object
in memory. This forces a re-render, and a new error, which forces
another re-render, ad eternum.

To aleviate this issue we had to extract and normalize the error message
and compare whether it had changed or not, and only then call
`setError`.

Now that we extract the `message` from the error object we don't have
that problem anymore. A `===` comparison against 2 strings comprised of
the same characters will always be true.