Rename invites to friends
The problem I saw with "invites" was that it looks at the feature from
the perspective of the system/developer, i.e. we are creating
invitations for others to use the system.

Instead I hope that this is more user-centered. You want to give access
to your friends to your music collection. It's not about the invitation
per-se, it's more about the people you want to invite.

The risk, however, is that the term "friend" has been abused by social
networks for the past 15 years, so it could potentially generate false
expectations that this is some kind of social network feature. The idea
is to keep usage as anonymous as possible, so none of the users will be
identifyiable really, and the owner will not be able to "spy" on what
the friends he/she invited are listening to.