Add list of orphan tracks
We admit that tracks will not always have an album associated. As a
consequence we need to check, in every place where we access
`track.album` whether an album actually exists. That's why this change
cannot be contained to a `src/routes/tracks` directory only and touches
so many other components as well.