Display invitation QR in a modal
Before we used to display the QR code right below the button. This had a
couple of issues: We disabled the button while the code was being
generated but hid it after the QR code was generated. To generate a new
code the user had to navigate away from the page and back. This is
terrible UX.

If we didn't hide it though when the user clicked it again the code
would regenerate but the change is very quick and subtle. The user could
miss the moment the code was replaced with the new one and wonder
whether anything changed at all.

Now when the user closes the modal the code is gone and the user is free
to regenerate it and they will get a new one, no way to get the old one
back. This really reflects the volatility of the invitations. There's no
way to get it back.

On top of that we can now add a small disclaimer that won't get lost in
the complexity of this view (which is quite cluttered as of today).