Refactor <Options> component
It was broken down into 2 independent components: a <Modal> component
that can be used standalone, and an <Options> component that simply adds
an icon trigger to the modal.

This is needed because we want to display the invitation's qr code in a
modal now, and we didn't have a standalone modal component so far.