Refactor useCreateInvite mutation
This commit extracts the generic parts of the mutation call so they can
be reused in future mutations. At the same time, the old code used to
rely on the apollo provider, that's why we used to have a call to
`getClient` in the `root.tsx` file.

Since we removed that call in a previous commit now this code stopped
working because it was not able to find the client in the context. This
refactor fixes that.