Swap `react-qr-code` for `qrcode`
The library `react-qr-code` was a very thin wrapper around a library
called `qr.js`. Since the react library was limited we decided to ditch
it and create our own wrapper on a more convenient library that could
generate more than just an svg file.

With `qrcode` we can generate a data URL which is very convenient for
caching, and <img /> elements are easier to style than svg files, that
are quite rigid.

Finally, the component to generate a QR code was split in 2:
<CachedQrCode /> and <VolatileQrCode />. The reason for this is that we
have 2 clear and incompatible use cases. On the servers page we need to
generate the code once and cache it, do not even try to render again
because we know that the data will not change between renders.

On the friends page, however, we need to display a different qr code
every time the user clicks the "create invitation" button. On top of
that, we also need to show the spinner until the new qr is generated and
we are sure that neither the input value nor the generated data will be
the same.

It was extremely difficult to merge these 2 use cases into one,
especially because the qr library that we use has a weird cache in which
consecutive qr generations show the old code for an instant and
immediately re-render with the new data. This is super annoying and very
difficult to detect or prevent, the best solution is to split the
component in 2 so we know what to expect from each.