Add a splash screen
There used to be this semi-FOUC while the header's background image and
the fonts were loading. This was very non-app-y. Maybe there is a better
way of doing this, with some service worker caching or something like
that, but for the time being we just show a splash screen while we
listen for the background and the fonts to fire their respective "ready"
events.