Remove invitation's validity from the UI
It made sense when the URL was supposed to be sent in an asynchronous
way, but now that we display a QR code (the implication being that the
code will be scanned synchronously) the invitation should be claimed
immediately.

The concept of validity for the backend still makes a lot of sense
though, we don't want to leave unclaimed invitations lying around that
could be leaked at some point. We want to invalidate them in a
reasonable amount of time.