Improve auth component
It was quite mess of inline code, which was mostly repeating code that
we already implemented in components like Spinner and ErrorMsg. Those
components were used instead and just a new one was created for the
secret-inputting-specific logic.

Also a spinner was used in the root.tsx file to show one unified spinner
throughout the whole loading process, instead of a string sometimes, and
a spinner some other times.