Use different domains for the invite and the app
Safari has a bug in which it ignores the cache and pragma headers for
the invitations. That means that the first time you load an invitation
it will work as expected, but if for some reason you have to load an
invitation again you will not be able to reach it, instead you will be
presented with the app asking you to enter your secret.

Neither http headers nor html meta tags solved this issue, so instead we
use this opportunity to do something even better: Use 2 completely
different domains for the invitation and the app. For the invite the
domain will be api gateway's, but only when you accept the invitation
you will be redirected to the actual app.

This had a couple of benefits, one of which I did not foresee but that I
really like. First, the user will not get the code for the app until the
invitation is claimed. This means not only less traffic, but also no
leaking of the app itself only by having access to an invitation (that
may be rejected by the user).

But more importantly, we can now have "in-app" instructions, for which
we can use the full power of preact to display in a very user-friendly
way with screenshots and all. It was very awkward to have these
instructions as read-only in plain HTML generated by `appendChild`-ing
and `innerText`-ing hardcoded strings.