Change default "active" behaviour for Link
By default the Link element will only show the active class if the path
matches the href exactly. In the new servers tab, however, we need to
keep the menu item highlighted not only when on the `/servers` url but
also in `/servers/add`.

To that end, we check if the path `beginsWith` the provided `href`
instead of doing a strict equality check.