Add track selection setting
Unlike all music players that I've ever used, this app revolves around
the idea of "album" rather than "track". Some albums have either long
intros or the first track is not a song at all, that's why in order to
skip that track but listen to the whole album, the default behaviour
when clicking on a track is to append all tracks, starting from the one
clicked.

I've noticed that this behaviour not intuitive for everyone, some people
look confused when after clicking on one track, the notification shows
that more than 1 track was added to the queue.

As an experiment I'm adding this feature, to select what the behaviour
of the app will be in this case. I'm worried that this only complicates
things even more because I'm not sure it's very clear what the setting
means and which parts of the app it affects.