Allow the spinner to be set an empty message
One of the reasons to hide the app behind a splash screen is that the
fonts are still loading. That's also why we set an empty spinner message
there (a `" "` to be exact) and show only the spinning record gif.

In this commit we fix the fact that the Spinner component requires a
non-null string message, which forced the splash screen to set `" "` as
a message.