Warn every time we touch localStorage
We did some efforts before to warn every time localStorage was accessed.
Those warnings, however, were not exhaustive, i.e. we warned on most of
localStorage accessess but not all.

On top of that we are also explicit about the operation we are doing,
rather than a generic "Touching localStorage" message that was not clear
about the operation that was taking place.