Delete custom Service Worker
I tried to introduce custom behaviour into the SW but given that:
a) I do not understand how the browser handles service workers,
b) I do not understand how preact-cli/webpack generates it, or how to
tap into that process,
c) I do not understand how `workbox` generates it, what the generated SW
does and how to modify it,

I will just put this aside for the time being and trust that the default
SW should be good enough.

What I was trying to accomplish was:
1. Prevent anything under `/assets/icons/svg/` to be cached
2. Check for updates and notify the user

However simple the first task seemed, I failed to find the way to
achieve that goal. There is documentation in the
`preact-cli-workbox-plugin` repository that I thought was exactly what I
was looking for but it didn't work:

https://github.com/prateekbh/preact-cli-workbox-plugin#generatesw
```
// in preact.config.js
const { generateSw } = require('preact-cli-workbox-plugin');
export default function(config, env, helpers) {
  return generateSw(config, helpers, {});
}
```

Regarding the check for updates, it's a mess. There's contradicting
documentation online about how browsers treat service workers, and even
the actual behaviour is confusing for me.

I was expecting the app to know that there was an update available when
there was a new SW downloaded (because it only changes when there's a
new build, and it will always change with each build). Some
documentation mentions that the browser will check if the new SW is
identical to the current one, and notify you there's a worker waiting to
be activated only if they differ. This is not true.

The browser will download the worker *every* time, regardless of whether
it changed or if it's bit-identical. It will not, however, activate it
automatically. This means that the system knows there is a new SW
waiting for activation but it doesn't know if it has changed or not.

Anyway, I tried to find a way to extend the default SW's behaviour but I
find myself fighting to understand the different layers that lead to the
generation of that file and how to tap into each one of them, and it's a
mess:
* preact-cli-workbox-plugin
* preact-cli
* webpack
* workbox

I give up.