Various fixes
When I deployed the `servers` branch in parallel to test the servers
functionality a bunch of issues became apparent that were overlooked
before. This branch started as a quickfix for those issues but it
quickly became a parallel `servers` development branch where I continued
the actual development of the feature.

After a few months of not touching the project, now that I come back to
it I find it almost impossible to break this down into atomic commits.

Here's a very important lesson kids: Do not have long-lived branches,
and commit quickly and often into atomic commits.