Rename `useQuery` hook to `useList`
The dependency for the hooks that get 1 item by id is as follows:
```
useGetEntity -> useGet -> use
```

While listing goes this way:
```
useListEntity -> useQuery -> use
```

It's amiguous that "query" is meant only for lists of items. One can
query 1 item as well. With this change the dependency chain is a lot
more clear:
```
useListEntity -> useList -> use
```