Fix issue with removed invites not disappearing
I am not sure why this fixes it tbh, but it does. The problem was that
when deleting an invite, the trash icon would disappear but the invite
would not.

My guess is that the <img> element with the `onClick` handler got
replaced by an empty <div /> preventing the callback (which would
execute the refetch) from ever being called.