Change the "add server" flow
It took me a while to understand all of cognito's concepts, like user
pool, app client, and identity provider. As a result, the implementation
was flawed. For example, The app client was being created with
`GenerateSecret=True`, which makes it impossible to use with web
clients.

Some not-so-minor improvements were made along the way, the most
important, imo, being that the app client is not created in the same
thread that the invitation is created. This meant that we had a bunch of
unused app clients lying around.

Now we create them on demand, when the other server requests the client
id, so we are sure that we only create resources that are being used.

This has its drawbacks as well tbh, because now the request to get the
client id takes much longer. Besides a very likely cold start, it has to
create the app client, on top of the DyanmoDB requests that it already
had to do. This led to the increase in timeout of the "register_server"
AppSync resolver.