Remove volume information from artist list
My first instinct was to add the number of volumes next to the album
name, I guess that influenced by how the mp3 for multi-volumed albums
are usually tagged, i.e. "Album Name (CD 1)".

But I realized that this is not necessary, and in fact, that's not how
artists distribute their music. The whole album has a name, and it
contains multiple volumes, but it's all published under one name.

So in this commit I remove the "(x2)" after the album name, and with
that the code becomes a lot simpler too. Honestly when I initially
introduced this feature I was sad how much logic now went into something
as simple as displaying the album name.