Add `path` attribute to tracks
This is used only for the orphan track list. Since some of those tracks
will lack all metadata (that's most likely why we couldn't figure out
which album they belong to) it's also likely that they won't even have a
title.

Showing the path that they were uploaded from will help the user
recognize what track that is. Once the track is assigned to an album,
the path is never displayed again.