Improve search behaviour
When a user searched, each new character pushed a new state into the
history. This meant that pressing the back button would "undo" the
search character by character. To make matters worse, the app was not
subscribed to the `hashchange` event so the user would not see anything
changing while pressing the back button, effectively hijacking it for as
many times as `onChange` events were triggered by the input.

In this new version we trigger 1 `hashchange` event when the first
character is pressed, but then call `history.replaceState` with each
subsequent `onChange` event. This means that the user can perform a
search, for example in the album list, then click on an album, and when
hitting back he/she would go first to the serach result, and a
subsequent click of the back button would clear the search and take them
to the album list again.