Unset the invite QR code once the modal is hidden
We need to wait for 250ms before clearing the QR code because otherwise
the spinner will show and immediately after the fadeout transition will
begin, making the spinner flash for a quarter of a second.

To avoid this we wait until the transition is over and we clear the QR
code right after.