Rename functions
These functions were named in a consistent way ("was_*") but the checks
they were performing was not similar in any way. Some were checking for
a value in the invite (stored in the DB) and some were looking at the
request to see into which state the invite had to be mutated.

Very confusing.