Suffix buckets with a 16-character random string
It was informed in a recent article that just by knowing the name of a
bucket a malicious actor could make the owner incur in high costs
without any way for the victim to protect themselves. In an attempt to
mitigate this risk we generate a random string for each bucket we
create. In theory the bucket name is never exposed until valid
credentials are provided, but just in case we add another layer of
security. For more information see:
https://medium.com/@maciej.pocwierz/how-an-empty-s3-bucket-can-make-your-aws-bill-explode-934a383cb8b1