Do not reload when changing servers
We used to `selectServer` by doing a full reload, calling
`window.location.href`. This is obviously problematic but we did it for
a reason: Once we call `setConf` the app won't re-render properly, the
only way to force a render and update the app with the information of
the new server was to reload the page. This is not ideal, mainly
because playback will be affected and that's unacceptable. We need to
fix that.