I don't know, man... *sigh*
This is quite an experimental app, I don't have a clear roadmap and in
fact I'm not even sure that the ideas I have are possible. On top of
that, I have enough frontend experience to build this app but I'm not a
frontend developer by no means. And this is quite a complex UI in which
the track has to keep playing while you navigate around (even to other
servers) and you can't re-render the whole app every time the playback
timer increases, among other challenging restrictions.

So while I embark in a new feature (some of which are quite huge) I find
tons of things to improve along the way, so I just do them. I haven't
been dilligent in creating atomic commits with every single unrelated
change, as much as I love git, it can be exhausting. Add to that the
fact that I don't work on this app for weeks or months at a time, and
when I come back I'm not even sure which change does what or belongs
where. I have to be honest, some changes are just "random", in the sense
that I try something in order to fix a problem, the problem is not fixed
but the code remains.

And it's fine, it's a side project. I do strive to have high standards
for my own projects, the kind of standards that are quickly disregarded
in any company. I want to lead by example. But this is also for fun,
it's not supposed to be yet another source of stress, quite the opposite
actually.

All this to say, yeah, this is a lot of unrelated code in one commit,
but when you are exploring, seeing what's possible, pushing the limits,
this happens. Until (or better: if) this project becomes more stable
then I'll be able to plan things ahead. But for now this is how things
are.

Thanks for coming to my TED talk :)